// mony_mvp/types/console.ts
export interface Account {
  id: string
  bank_code?: string
  bank_name: string
  account_type: AccountType
  balance: number
  account_number?: string | null
  last_updated?: string | null
}

export enum AccountType {
  CHECKING = 'CHECKING',
  SAVINGS = 'SAVINGS',
  INVESTMENT = 'INVESTMENT',
  CREDIT = 'CREDIT',
}